// prototipo-angular/src/app/core/models/models.ts
export interface SoftwareCatalogo {
  codigo: string;
  nombre: string;
  categoria: string;
  versionesPermitidas: string[];
  versionVigente: string;
  aplicaCPU: boolean;
  aplicaLaptop: boolean;
  aplicaF0288: boolean;
  aplicaF0302: boolean;
  activo: boolean;
  observacion: string;
}